Login Completed
Checks to see if the username and password are correct. Using switch to
check which user is logged in.

Moved all other classes to a package called "support"

Might just create the user object and return it to the
login controller instead, but this is how you work the database atleast