Login now creates the User object.
I changed the login verification where it now creates the User object
with respect to the user type.

I also made Schedule a Super class and extended two types of schedules
one for the Student and one for the Professor. I the variable from the
User class "Credits" and put that in the Student Schedule Class since
only the student will need to have number credits in their schedule.

We are storing the Schedule in a database as their course numbers
delimited by a whitespace.

When we need to find courses we just read the schedule of the specific
user, using the split function we can create a String[] of all of the
courses. After that is done we can can use the parse.Integer function to
convert the strings into integers. Then we can look each course up by
their courseid.

Let me know if you have any questions.